very good early version
