--- CuriousRL/algorithm/__init__.py
@@ -1,7 +1,0 @@
-from .ilqr_solver.basic_ilqr import BasiciLQR
-from .ilqr_solver.log_barrier_ilqr import LogBarrieriLQR
-
-
-
-# from .ilqr_solver.advanced_ilqr import NNiLQR
-
