+++ CuriousRL/data/__init__.py
@@ -1,0 +1,3 @@
+    """
+        We borrow the data time from pytorch to construct the data module
+    """
