+++ CuriousRL/scenario/dynamic_model/car_parking.py
@@ -2,6 +2,9 @@ import numpy as np
 import sympy as sp
 from .dynamic_model import DynamicModelWrapper
 from CuriousRL.utils.Logger import logger
+import matplotlib.pyplot as plt
+import matplotlib.patches as patches
+import matplotlib as mpl
 
 class CarParking(DynamicModelWrapper):
     """ In this example, the vehicle packing at 0, 0, heading to the top\\
@@ -40,13 +43,11 @@ class CarParking(DynamicModelWrapper):
             else: 
                 add_param_obj[tau] = np.asarray((0.,1.))
 
-        C_matrix = np.diag([0.,1.,1.,1.,10.,10.])
-        r_vector = np.asarray([0.,-10.,0.,8.,0.,0.])
         def Huber_fun(x, p):
             return sp.sqrt((x**2)+(p**2)) - p
         runing_obj = Huber_fun(x_u_var[0], 0.01) + Huber_fun(x_u_var[1], 0.01) + Huber_fun(x_u_var[2], 0.01)
         terminal_obj = 1000*Huber_fun(x_u_var[0], 0.01) + 1000*Huber_fun(x_u_var[1], 0.01) + 1000*Huber_fun(x_u_var[2], 0.01) + 100*Huber_fun(x_u_var[3], 0.01)
-        input_obj = x_u_var[4]**2+x_u_var[5]**2
+        input_obj = x_u_var[4]**2 + x_u_var[5]**2
         obj_fun = switch_var[0] * runing_obj + switch_var[1]*terminal_obj + input_obj
         super().__init__(   dynamic_function=dynamic_function, 
                             x_u_var = x_u_var, 
@@ -67,23 +68,18 @@ class CarParking(DynamicModelWrapper):
             no_iter : int
                 The number of iteration to play the animation
         """
-        import matplotlib.pyplot as plt
-        import matplotlib.patches as patches
-        import matplotlib as mpl
+        fig, ax, current_player_id = super().create_plot(figsize=(5, 5), xlim=(-5,10), ylim=(-7.5,7.5))
         trajectory = np.asarray(logger.read_from_json(logger_folder, no_iter)["trajectory"])
-        fig = plt.figure(figsize=(5, 5)) 
-        ax = fig.add_subplot(111) 
         car = patches.FancyBboxPatch((0, 0), 3, 2, "round,pad=0.2")
         car.set_color('C0')
         ax.add_patch(car)
-        ax.axis('equal')
-        plt.xlim((-5, 10))
-        plt.ylim((-7, 7))
         plt.plot(trajectory[:,0], trajectory[:,1],'C1')
         plt.plot([-1,-1], [10,-10],'C2')
         plt.plot([-1,5], [2,2],'C2')
         plt.plot([-1,5], [-2,-2],'C2')
         for i in range(self.get_T()):
+            if self.check_interrupted(current_player_id): # if this player is not interrupted
+                break
             angle = trajectory[i,2,0]
             t_start = ax.transData
             x = trajectory[i,0,0] + 1*np.sin(angle)
@@ -95,6 +91,5 @@ class CarParking(DynamicModelWrapper):
             t_end = t_start + t
             car.set_transform(t_end)
             fig.canvas.blit(fig.bbox)
-            plt.pause(0.01)
-        plt.show()
+            plt.pause(0.001)
