+++ CuriousRL/scenario/dynamic_model/cart_pole_swingup2.py
@@ -3,6 +3,9 @@ import numpy as np
 import sympy as sp
 from .dynamic_model import DynamicModelWrapper
 from CuriousRL.utils.Logger import logger
+import matplotlib.pyplot as plt
+import matplotlib.patches as patches
+import matplotlib as mpl
 
 class CartPoleSwingUp2(DynamicModelWrapper):
     """ In this example, the cartpole system is static at 0, 0, heading to the postive direction of the y axis\\
@@ -65,22 +68,17 @@ class CartPoleSwingUp2(DynamicModelWrapper):
             no_iter : int
                 The number of iteration to play the animation
         """
-        import matplotlib.pyplot as plt
-        import matplotlib.patches as patches
-        import matplotlib as mpl
+        fig, ax, current_player_id = super().create_plot(figsize=(5, 2), xlim=(-5, 5), ylim=(-1,1))
         trajectory = np.asarray(logger.read_from_json(logger_folder, no_iter)["trajectory"])
-        fig = plt.figure(figsize=(5, 2)) 
-        ax = fig.add_subplot(111) 
         cart = patches.FancyBboxPatch((0, -0.1), 0.4, 0.2, "round,pad=0.02")
         cart.set_color('C0')
         pole = patches.FancyBboxPatch((0, 0), 0.04, 0.5, "round,pad=0.02")
         pole.set_color('C1')
         ax.add_patch(cart)
         ax.add_patch(pole)
-        ax.axis('equal')
-        plt.xlim((-5, 5))
-        plt.ylim((-1, 1))
         for i in range(self.get_T()):
+            if self.check_interrupted(current_player_id): # if this player is not interrupted
+                break
             angle = np.arctan2(trajectory[i,2,0], trajectory[i,3,0])
             t_start = ax.transData
             x = trajectory[i,0,0]-0.02*np.cos(angle)
@@ -94,4 +92,3 @@ class CartPoleSwingUp2(DynamicModelWrapper):
             cart.set_x(trajectory[i,0]-0.2)
             fig.canvas.blit(fig.bbox)
             plt.pause(0.01)
-        plt.show()
