+++ test.py
@@ -1,15 +1,14 @@
 #%%
-from CuriousRL.algorithm import BasiciLQR, LogBarrieriLQR
+from CuriousRL.algorithm.ilqr_solver import BasiciLQR, LogBarrieriLQR
 from CuriousRL.scenario.dynamic_model import CartPoleSwingUp1, CartPoleSwingUp2, \
                                             VehicleTracking, CarParking,\
                                             RoboticArmTracking, TwoLinkPlanarManipulator, \
                                             ThreeLinkPlanarManipulator
-from CuriousRL import ProblemBuilderClass
 
 if __name__ == "__main__":
-    # scenario = RoboticArmTracking(is_with_constraints=True)
-    scenario = TwoLinkPlanarManipulator()
-    algo = LogBarrieriLQR(max_iter = 1000)
-    ProblemBuilderClass(scenario , algo).learn(logger_folder="test2")
+    scenario = RoboticArmTracking()
+    algo = BasiciLQR()
+    algo.init(scenario,is_save_json=False, is_use_logger= False).solve()
     scenario.play()
+
 # %%
